requirements.txt
openai-whisper
torch
torchaudio
imageio[ffmpeg]
pyinstaller
ttkbootstrap
torchaudio --index-url https://download.pytorch.org/whl/cu118
